# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZeroRoot API",
    "version": "1.0.0",
    "description": "Secure identity, RBAC, tenant isolation, scanning, chat, and central AI gateway API."
  },
...
  "paths": {
...
    "/auth/{portal}/login": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "portal",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/auth/sessions/{id}": {
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/v1"
    }
  ]
}